Delete Billabong.ttf
Please be advised that the font Billabong, which has been detected at this website, is subject to an End User License Agreement which specifically prohibits free downloading or file sharing. The software is neither freeware nor shareware and can only be obtained by purchasing a license from an authorised agent of Type Associates, the creator.